+++ packages/anon-aadhaar-pcd/src/utils.ts
@@ -261,3 +261,114 @@ export function decompressByteArray(byteArray: Uint8Array) {
   const decompressedArray = pako.inflate(byteArray)
   return decompressedArray
 }
+
+export const enum SELECTOR_ID {
+  emailOrPhone = 0,
+  referenceId,
+  name,
+  dob,
+  gender,
+  cafeof,
+  district,
+  landmark,
+  house,
+  location,
+  pinCode,
+  postOffice,
+  state,
+  stress,
+  subDistrict,
+  VTC,
+}
+
+export class SelectorBuilder {
+  selector: number[]
+
+  constructor() {
+    this.selector = new Array(16).fill(0)
+  }
+
+  selectEmailOrPhone() {
+    this.selector[0] = 1
+    return this
+  }
+
+  selectReferenceId() {
+    this.selector[1] = 1
+    return this
+  }
+  selectName() {
+    this.selector[2] = 1
+    return this
+  }
+  selectDoB() {
+    this.selector[3] = 1
+    return this
+  }
+  selectGender() {
+    this.selector[4] = 1
+    return this
+  }
+  selectCafeof() {
+    this.selector[5] = 1
+    return this
+  }
+  selectDistrict() {
+    this.selector[6] = 1
+    return this
+  }
+  selectLandmark() {
+    this.selector[7] = 1
+    return this
+  }
+  selectHouse() {
+    this.selector[8] = 1
+    return this
+  }
+  selectLocation() {
+    this.selector[9] = 1
+    return this
+  }
+  selectPinCode() {
+    this.selector[10] = 1
+    return this
+  }
+  selectPostOffice() {
+    this.selector[11] = 1
+    return this
+  }
+  selectState() {
+    this.selector[12] = 1
+    return this
+  }
+  selectStress() {
+    this.selector[13] = 1
+    return this
+  }
+  selectSubDistrict() {
+    this.selector[14] = 1
+    return this
+  }
+  selectVTC() {
+    this.selector[15] = 1
+    return this
+  }
+
+  build() {
+    return this.selector
+  }
+}
+
+export function readData(data: number[], index: number) {
+  let count = 0
+  let start = 0
+  let end = data.indexOf(255, start)
+
+  while (count != index) {
+    start = end + 1
+    end = data.indexOf(255, start)
+    count++
+  }
+
+  return data.slice(start, end)
+}
