+++ packages/circuits/circuits/nullifier.circom
@@ -1,0 +1,4 @@
+// nullifier = poseidon(LAST 4 DIGITS + NAME + DATE OF BIRTH + name) 
+// referenceId = last 4 digits of aadhaar number + 'DDMMYYYYHHMMSSsss';
+// DDMMYYYYHHMMSSsss < 30 mins;
+
